pyproject.toml
[project]
name = "arbitration-api"
version = "0.1.0"
description = "Arbitration API - Manage legal arbitration cases with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.40.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
